Make it possible to run only one specified suite of tests
use: time docker compose exec api deno test -A --unstable --cached-only tests/smoke_26_session.ts
