+++ backend/tests/smoke_25_subject.ts
@@ -3,8 +3,13 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 // TODO: add bad requests
+import { basename } from "../deps.ts";
 import { assert, assertEquals } from "../test_deps.ts";
-import { endpointFactory, RoleTestContext } from "./smoke_mod.ts";
+import {
+  endpointFactory,
+  registerRoleTest,
+  RoleTestContext,
+} from "./smoke_mod.ts";
 import { data } from "./testdata/config.ts";
 
 interface DataEndpoint {
@@ -274,3 +279,5 @@ export async function initSubjectTests(
     );
   });
 }
+
+registerRoleTest(basename(import.meta.url), initSubjectTests);
